Add support for Swift Package Manager
This closes #1, closes #3, closes #5
