+++ Apexy.podspec
@@ -9,21 +9,17 @@ Pod::Spec.new do |s|
 
   s.ios.deployment_target = "10.0"
   s.tvos.deployment_target = "10.0"
-  s.osx.deployment_target = "10.10"
+  s.osx.deployment_target = "10.12"
   s.watchos.deployment_target = "4.0"
 
   s.swift_version = "5.2"
-  s.source_files  = "Source/Apexy/**/*.swift"
+  s.source_files  = "Sources/Apexy/**/*.swift"
   
   s.dependency "Alamofire", '~>5.0'
 
   s.subspec 'RxSwift' do |sp|
-    sp.source_files = 'Source/Apexy+RxSwift/*.swift'
+    sp.source_files = 'Sources/Apexy+RxSwift/*.swift'
     sp.dependency "RxSwift"
-  end
-
-  s.subspec 'Combine' do |sp|
-    sp.source_files = 'Source/Apexy+Combine/*.swift'
   end
 
   s.default_subspecs = :none
