+++ README.md
@@ -1,5 +1,9 @@
 # Apexy
 
+[![SPM compatible](https://img.shields.io/badge/spm-compatible-brightgreen.svg?style=flat)](https://swift.org/package-manager)
+[![Swift 5.2](https://img.shields.io/badge/swift-5.2-red.svg?style=flat)](https://developer.apple.com/swift)
+[![GitHub license](https://img.shields.io/badge/license-MIT-lightgrey.svg)](https://github.com/RedMadRobot/api-client-ios/blob/master/LICENSE)
+
 The library for organizing a network layer in a project.
 
 - The library doesn't depend on the implementation of how to work with the network. You can use `URLSession` or `Alamofire`.
