+++ README.ru.md
@@ -1,5 +1,9 @@
 # Client
 
+[![SPM compatible](https://img.shields.io/badge/spm-compatible-brightgreen.svg?style=flat)](https://swift.org/package-manager)
+[![Swift 5.2](https://img.shields.io/badge/swift-5.2-red.svg?style=flat)](https://developer.apple.com/swift)
+[![GitHub license](https://img.shields.io/badge/license-MIT-lightgrey.svg)](https://github.com/RedMadRobot/api-client-ios/blob/master/LICENSE)
+
 Библиотека для организации сетевого слоя в проекте.
 
 - Не зависит от реализации работы с сетью. Может быть использована `URLSession` или `Alamofire`.
