--- Source/Apexy/Supporting Files/Apexy.h
@@ -1,14 +1,0 @@
-//
-//  Apexy.h
-//
-//  Created by Anton Glezman on 18/06/2019.
-//  Copyright © 2019 RedMadRobot. All rights reserved.
-//
-
-#import <UIKit/UIKit.h>
-
-//! Project version number for Apexy.
-FOUNDATION_EXPORT double ApexyVersionNumber;
-
-//! Project version string for Apexy.
-FOUNDATION_EXPORT const unsigned char ApexyVersionString[];
