--- Source/Podfile
@@ -1,8 +1,0 @@
-platform :ios, '10.0'
-
-target 'Apexy' do
-  use_frameworks!
-
-  pod 'Alamofire'
-
-end
